requirements.txt
# Kafka
kafka-python-ng==2.2.2
confluent-kafka[avro]==2.3.0
avro-python3==1.10.2
fastavro==1.9.4

# API & HTTP
requests==2.31.0
aiohttp==3.9.1

# Utils
six==1.16.0
python-dateutil==2.8.2

# Monitoring (optionnel pour cette phase)
prometheus-client==0.19.0

# Data processing
pandas==2.1.4
numpy==1.26.3
